[week4-task2] page title을 추출하고 page의 소제목과 링크 연결하는 기능 구현
