+++ src/PageTitlesContainer.jsx
@@ -1,21 +1,37 @@
 import React from 'react';
 
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import PageTitles from './PageTitles';
 
-import { page1Titles } from './../fixtures';
-
 export default function PageTitlesContainer() {
-  // const dispatch = useDispatch();
 
-  // const courses = useSelector((state) => state.book.courses);
+  const opendPage = useSelector((state) => state.book.opendPage);
 
-  
+  function extractSubTitlesInPage(pageText) {
+    const titles = pageText.match(/##\D+?\n/gm) || [];
+
+    return titles.map(title => {
+      return title.replace(/(##\s|##)/g, '')
+                  .replace(/\n/g, '');
+    });
+  }
+
+  function getPageTitles(subTitlesInPage) {
+    return subTitlesInPage.map((title, index) => ({
+      id: index,
+      title: title,
+      cssId: `#${title.replace(/\s/g, '-')}`
+    }))
+  };
 
   return (
     <>
-      <PageTitles titles={page1Titles} />
+      <PageTitles 
+        titles={getPageTitles(
+            extractSubTitlesInPage(opendPage.text)
+          )}
+      />
     </>
   );
 }
